+++ SkyModel/Sky/ClassSM.py
@@ -37,10 +37,10 @@ class ClassSM():
 #         if "instance" in str(type(infile)):
 #             ClusterCat=infile.ClusterCat.copy()
 #             NN=ClusterCat.shape[0]
-#             Cat=np.zeros((NN,),dtype=[('Name','|S200'),('ra',np.float),('dec',np.float),('Sref',np.float),('I',np.float),('Q',np.float),\
-#                                                        ('U',np.float),('V',np.float),('RefFreq',np.float),('alpha',np.float),('ESref',np.float),\
-#                                                        ('Ealpha',np.float),('kill',np.int),('Cluster',np.int),('Type',np.int),('Gmin',np.float),\
-#                                                        ('Gmaj',np.float),('Gangle',np.float),("Select",np.int),('l',np.float),('m',np.float),
+#             Cat=np.zeros((NN,),dtype=[('Name','|S200'),('ra',float),('dec',float),('Sref',float),('I',float),('Q',float),\
+#                                                        ('U',float),('V',float),('RefFreq',float),('alpha',float),('ESref',float),\
+#                                                        ('Ealpha',float),('kill',int),('Cluster',int),('Type',int),('Gmin',float),\
+#                                                        ('Gmaj',float),('Gangle',float),("Select",int),('l',float),('m',float),
 #                                                        ("Exclude",bool)])
 #             Cat=Cat.view(np.recarray)
 #             Cat.RefFreq=1.
@@ -435,7 +435,7 @@ class ClassSM():
         self.SourceCat.Select=M.Start()
 
     def BuildClusterCat(self):
-        ClusterCat=np.zeros((len(self.Dirs),),dtype=[('Name','|S200'),('ra',np.float),('dec',np.float),('SumI',np.float),("Cluster",int)])
+        ClusterCat=np.zeros((len(self.Dirs),),dtype=[('Name','|S200'),('ra',float),('dec',float),('SumI',float),("Cluster",int)])
         ClusterCat=ClusterCat.view(np.recarray)
         icat=0
         for d in self.Dirs:
